using 2 config files
updating to use 2 config files. This allows user/machine specific data to be stored in user.config and the common data needed for running to be stored in run.config. By doing this, we can commit run.config without worrying that it will affect someone else's environment. They will still be able to get the default ports, volumes, etc.